Start encoding assembly syntax. No tests.
This is very repetitive, but the repetition may be inherent to the task.
In an ISA there will inevitably be many things that are similar but all
a little bit different in different ways, and the task at hand is to
encode the entire syntax, which is not a tiny amount of information.