Add addattr, hasattr.
This is a little interesting because they are HOFs implemented in OCaml,
rather than in the language that is being interpreted.

I'm not sure I understand why I do not need another pair of brackets []
around the selected branch of the "if" in stdlib. This might be a bug,
or maybe I am just tired.