Get prologue and epilogue to work.
The dynamic checks involved uncovered one or two problems and were a
major reason why this was difficult.

In the process it was necessary to make isLam? work for lams created
by builtins. This requires special handling because those do not have
parameter lists.

Also change several of the ranges reported, and change exprToString to
include attributes. Hopefully the reported ranges are not more wrong
than they were before.